Moved equals() from *Quantity to JavaNumberQuantity
